+++ spring-data-r2dbc-sqlfile-example/src/main/java/com/example/MsgRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.r2dbc.repository.R2dbcRepository;
 
 import reactor.core.publisher.Flux;
 
-public interface MsgRepository extends R2dbcRepository<Msg, Integer> {
+public interface MsgRepository extends R2dbcRepository<Msg, Long> {
 
     @SqlFile
     Flux<Msg> findAllMsgs();
